Add doxygen file to automagically create documentation.
I have added a doxygen.config file. You can use doxygen command
> doxygen doxygen.config
to create the documentation. It will be stored in the directory
doc/ . I also added this directory to .gitignore. In order to
create nice graphs (class inheritance for example) you need the
dot tool which is part of the graphviz package.