Use CMake for the build.
I have written a CMakeLists.txt so that the project can now be
compiled and installed using CMake. This has several advantages:

1) You can now use "make install" to install the library and the
   python module (no need for sys.path.append() commands in the
   python scripts anymore).
2) Easy out-of-source build
3) First step towards cross-platform build
4) ...

The typical CMake workflow is:

> mkdir build/ && cd build/
> cmake ..
> make

To activate building the python library type

> cmake -DPYTHON=1 ..

and to activate debug output type

> cmake -DDEBUG=1 ..

and, of cause, you can combine the two. If you want to install
the library and python module, type

> sudo make install

It's just this easy.