GPUArrayDevice: Allow calling allocate() when n == 0
Now, allocate will check if n == 0 and if true not allocate anything
and set the pointer to NULL. This means we can remove the constructor
GPUArrayDevice() and make n_ default to 0 in GPUArrayDevice(int n_).
Furthermore, we can remove checks that were inteded to make sure that
GPUArrayDevice(int n_) is not called with n_ = 0. The checks weren't
consistent anyhow.