Bugfix: InitializeAtoms::populateRand() did not check for overlap         across boundary.
When InitializeAtoms::populateRand() is called with distMin set,
the function checks if Atoms overlap. It did, however, not take the
boundaries into account. Thus, it was allowed to place an atom at
the very low edge and another at the very high edge of the box even
though they could overlap across the boundary. This issue is now
fixed.