CRITICAL: Fix Error in IntegraterRelax.cu
Just like in the other case, a float4 addition scrambled the atom
type. These errors are really tricky and difficult to spot. We
really should think about using a different approach. In
particular, are float4s such a big advantage? Especially since we
have to call a lot of make_float3(float4)s.