Remove function CopyFromOther
The function must not be called when memory is not deallocated
first and should only be used by the move constructor and move
assignment. Despite its name, it must not be called in the copy
constuctor or copy assignement operator.

I removed the function, and copy the member variables in the move
constructor/assignment instead. This is clearer and avoids that
copyFromOther is accendentally called where it shouldn't.