Do not allocate memory for passed nullpointer in get()
I follow the philosophy that memory should be freed and allocated
in the same spot. Thus, allocating memory in GPUArrayDeviceTex and
hoping that it will be freed somewhere else can easily lead to
memory leaks.

In the case of GPUArrayDeviceTex the change is trivial, since the
function never got passed a nullptr.